+++ package.json
@@ -6,7 +6,8 @@
     "dev": "ncc build ./index.ts -w -o dist/",
     "prerelease": "rimraf ./dist/",
     "release": "ncc build ./index.ts -o ./dist/ --minify --no-cache --no-source-map-register",
-    "prepublish": "yarn release"
+    "prepublish": "yarn release",
+    "test": "node --experimental-vm-modules node_modules/.bin/jest"
   },
   "files": [
     "dist"
@@ -24,9 +25,13 @@
   },
   "homepage": "https://github.com/Dhaiwat10/priv-key-precommit#readme",
   "devDependencies": {
+    "@types/jest": "^27.5.0",
     "@types/node": "^17.0.31",
     "@vercel/ncc": "^0.33.4",
+    "jest": "^28.1.0",
+    "jest-environment-node": "^28.1.0",
     "rimraf": "^3.0.2",
+    "ts-jest": "^28.0.2",
     "typescript": "^4.6.4"
   },
   "dependencies": {
