+++ readme.md
@@ -1,9 +1,2 @@
-- [x] get a list of all the files in the current directory
-- [x] skip files inside of .gitignore and other known stuff like node_modules
-- [x] loop through all the files
-- [ ] read the file and tokenize the contents
-- [ ] check for a private key
-- [ ] keep a running count of the number of private keys found and where they are
-- [ ] print out the whereabouts of all the private keys found (DONT print the keys themselves), plus throw an error if there are any to make sure husky fails
+# privy-key-precommit
 
-- [ ] write test cases
