+++ tests/index.test.ts
@@ -1,0 +1,7 @@
+jest.useFakeTimers();
+
+import { run } from '../utils';
+
+test('test', () => {
+  expect(run());
+});
